+++ platform/src/common/components/SideBar/AuthenticatedSidebar.jsx
@@ -11,6 +11,7 @@ import BarChartIcon from '@/icons/SideBar/BarChartIcon';
 import CollocateIcon from '@/icons/SideBar/CollocateIcon';
 import OrganizationDropdown from '../Dropdowns/OrganizationDropdown';
 import { checkAccess } from '@/core/utils/protectedRoute';
+import CollapsedSidebar from './CollapsedSidebar';
 
 const AuthenticatedSideBar = ({ toggleDrawer, setToggleDrawer, collapsed, setCollapsed }) => {
   const sideBarDisplayStyle = toggleDrawer ? 'flex fixed left-0 z-50' : 'hidden';
@@ -100,30 +101,7 @@ const AuthenticatedSideBar = ({ toggleDrawer, setToggleDrawer, collapsed, setCol
       </div>
     </div>
   ) : (
-    <div className='w-[88px]'>
-      <div
-        className={`${
-          size.width >= 1024 ? 'flex' : sideBarDisplayStyle
-        } bg-white h-[calc(100vh)] lg:relative flex-col justify-between overflow-y-auto border-t-0 border-r-[1px] border-r-grey-750 scrollbar-thin scrollbar-thumb-gray-800 scrollbar-track-gray-200`}>
-        <div className='flex flex-col items-center justify-center'>
-          <div className='p-4 justify-between items-center flex'>
-            <AirqoLogo className='w-[46.56px] h-8 flex flex-col flex-1' />
-          </div>
-          <div className='mt-3 mx-2'>
-            <SidebarIconItem IconComponent={HomeIcon} navPath='/Home' />
-            <SidebarIconItem IconComponent={BarChartIcon} navPath='/analytics' />
-            <hr className='my-3 h-[0.5px] bg-grey-150' />
-            {checkAccess('CREATE_UPDATE_AND_DELETE_NETWORK_DEVICES') && (
-              <SidebarIconItem IconComponent={CollocateIcon} navPath='/collocation/overview' />
-            )}
-            <SidebarIconItem IconComponent={WorldIcon} navPath='/map' />
-          </div>
-        </div>
-        <div className='mx-2 mb-2 flex items-center justify-center'>
-          <SidebarIconItem IconComponent={SettingsIcon} navPath='/settings' />
-        </div>
-      </div>
-    </div>
+    <CollapsedSidebar />
   );
 };
